backend/requirements.txt
aiofiles==24.1.0
fastapi==0.111.0
openai==1.35.13
pydantic_settings==2.3.4
pydub==0.25.1
uvicorn==0.30.1
python-multipart==0.0.9